Добавлены шаблоны и главный сценарий
