+++ templates/data.php
@@ -1,0 +1,54 @@
+﻿                    <?php
+                    $mainNavigation = [  
+                        "Все",
+                        "Входящие",
+                        "Учеба",
+                        "Работа",
+                        "Домашние дела",
+                        "Авто"
+                    ];
+                    
+                    $tasks = [
+                        [
+                            "task" => "Собеседование в IT компании",
+                            "deadline" => "01.06.2017",
+                            "type" => "Работа",
+                            "completed " => false
+                        ],
+                        [
+                            "task" => "Выполнить тестовое задание",
+                            "deadline" => "25.05.2018",
+                            "type" => "Работа",
+                            "completed" => false
+                        ],
+
+                        [
+                            "task" => "Сделать задание первого раздела",
+                            "deadline" => "21.04.2018",
+                            "type" => "Учеба",
+                            "completed" => true
+                        ],
+                        [
+                            "task" => "Встреча с другом",
+                            "deadline" => "22.04.2018",
+                            "type" => "Входящие",
+                            "completed" => false
+                        ],
+
+                        [
+                            "task" => "Купить корм для кота",
+                            "deadline" => "",
+                            "type" => "Домашние дела",
+                            "completed" => false
+                        ],
+                        [
+                            "task" => "Заказать пиццу",
+                            "deadline" => "",
+                            "type" => "Домашние дела",
+                            "completed" => false
+                        ]
+                    ];
+                    
+                    ?>
+
+
